move requirements to a separated file
